% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-DataFrame.R
\name{aggregatesFor}
\alias{aggregatesFor}
\title{Returns a list of aggegration functions for the given
variant caller}
\usage{
aggregatesFor(caller)
}
\description{
Returns a list of aggegration functions for the given
variant caller
}

