% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{align_breakpoints}
\alias{align_breakpoints}
\title{Adjusts the nominal position of a breakpoints}
\usage{
align_breakpoints(vcf, align = c("centre"),
  is_higher_breakend = names(vcf) < info(vcf)$PARID)
}
\description{
Adjusts the nominal position of a breakpoints
}
