% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{bedpe2breakpointgr}
\alias{bedpe2breakpointgr}
\title{Loads a breakpoint GRanges from a BEDPE file}
\usage{
bedpe2breakpointgr(file, placeholderName = "bedpe")
}
\arguments{
\item{file}{BEDPE file}

\item{placeholderName}{prefix to use to ensure ids are unique}
}
\value{
breakpoint GRanges object
}
\description{
Loads a breakpoint GRanges from a BEDPE file
}

