% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{blastHomology}
\alias{blastHomology}
\title{Identifies breakpoint sequences with signficant homology to BLAST database
sequences. Apparent breakpoints containing such sequence are better explained
by the sequence from the BLAST database such as by alternate assemblies.}
\usage{
blastHomology(gr, ref, db, anchorLength = 150)
}
\description{
Identifies breakpoint sequences with signficant homology to BLAST database
sequences. Apparent breakpoints containing such sequence are better explained
by the sequence from the BLAST database such as by alternate assemblies.
}
\details{
See https://github.com/mhahsler/rBLAST for rBLAST package installation
instructions
Download and install the package from AppVeyor or install via install_github("mhahsler/rBLAST") (requires the R package devtools)
}

