% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{breakendRanges}
\alias{breakendRanges}
\title{Extracts unpartnered breakend structural variants as a GRanges}
\usage{
breakendRanges(x, ...)
}
\description{
See Section 5.4.9 of https://samtools.github.io/hts-specs/VCFv4.3.pdf for details.
}
