% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{breakpointgr2pairs}
\alias{breakpointgr2pairs}
\title{Converts a breakpoint GRanges object to a Pairs object}
\usage{
breakpointgr2pairs(bpgr, writeQualAsScore = TRUE, writeName = TRUE,
  firstInPair = function(gr) {     seq_along(gr) < match(gr$partner,
  names(gr)) })
}
\arguments{
\item{bpgr}{breakpoint GRanges object}

\item{writeQualAsScore}{write the breakpoint GRanges QUAL field as the score
fields for compatibility with BEDPE rtracklayer export}

\item{writeName}{write the breakpoint GRanges QUAL field as the score
fields for compatibility with BEDPE rtracklayer export}

\item{firstInPair}{function that returns TRUE for breakends that are considered
the first in the pair, and FALSE for the second in pair breakend. By default,
the first in the pair is the breakend with the lower ordinal in the breakpoint
GRanges object.}
}
\value{
Pairs GRanges object suitable for export to BEDPE by rtracklayer
}
\description{
Converts a breakpoint GRanges object to a Pairs object
}
\examples{
vcf.file <- system.file("extdata", "gridss.vcf", package = "StructuralVariantAnnotation")
bpgr <- breakpointRanges(VariantAnnotation::readVcf(vcf.file))
pairgr <- breakpointgr2pairs(bpgr)
rtracklayer::export(pairgr, con="example.bedpe")
}
