% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{breakpointGRangesToVCF}
\alias{breakpointGRangesToVCF}
\title{Converts the given breakpoint GRanges object to VCF format in breakend
notation.}
\usage{
breakpointGRangesToVCF(gr)
}
\arguments{
\item{gr}{breakpoint GRanges object. Can contain both breakpoint and single breakend SV records}
}
\description{
Converts the given breakpoint GRanges object to VCF format in breakend
notation.
}
