% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{breakpointSequence}
\alias{breakpointSequence}
\title{Extracts the breakpoint sequence.}
\usage{
breakpointSequence(gr, ref, anchoredBases, remoteBases = anchoredBases)
}
\arguments{
\item{gr}{breakpoint GRanges}

\item{ref}{Reference BSgenome}

\item{anchoredBases}{Number of bases leading into breakpoint to extract}

\item{remoteBases}{Number of bases from other side of breakpoint to extract}
}
\description{
Extracts the breakpoint sequence.
}
\details{
The sequence is the sequenced traversed from the reference anchor bases
to the breakpoint. For backward (-) breakpoints, this corresponds to the
reverse compliment of the reference sequence bases.
}

