% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-DataFrame.R
\name{calculate_aggregates}
\alias{calculate_aggregates}
\title{Unpacks the columns of the given data frame}
\usage{
calculate_aggregates(x, caller = "gridss", F = aggregatesFor(caller))
}
\description{
Unpacks the columns of the given data frame
}

