% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{.breakendRanges}
\alias{.breakendRanges}
\title{Function for extracting the unpartnered structural variants as a GRanges}
\usage{
.breakendRanges(vcf, nominalPosition = FALSE,
  placeholderName = "svrecord", suffix = "_bp", info_columns = NULL)
}
\arguments{
\item{vcf}{VCF object}

\item{nominalPosition}{Determines whether to call the variant at the
nominal VCF position, or to call the confidence interval (incorporating
any homology present).}

\item{placeholderName}{variant name prefix to assign to unnamed variants}

\item{suffix}{suffix to append}

\item{info_columns}{VCF INFO columns to include in the GRanges object}
}
\description{
Function for extracting the unpartnered structural variants as a GRanges
}
