% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{.breakpointRanges}
\alias{.breakpointRanges}
\title{Function for extracting structural variants as GRanges}
\usage{
.breakpointRanges(vcf, nominalPosition = FALSE,
  placeholderName = "svrecord", suffix = "_bp", info_columns = NULL,
  unpartneredBreakends = FALSE)
}
\arguments{
\item{vcf}{a VCF object}

\item{nominalPosition}{Determines whether to call the variant at the
nominal VCF position, or to call the confidence interval (incorporating
any homology present).}

\item{placeholderName}{variant name prefix to assign to unnamed variants}

\item{suffix}{suffix to append}

\item{info_columns}{VCF INFO columns to include in the GRanges object}
}
\description{
Function for extracting structural variants as GRanges
}
