% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDetect.R
\name{.combineMatchingTranscripts}
\alias{.combineMatchingTranscripts}
\title{Combining matching transcripts}
\usage{
.combineMatchingTranscripts(gr, names)
}
\arguments{
\item{gr}{A GRanges object}

\item{names}{A vector of granges names.}
}
\value{
A list of vectors. Each vector is named with the name of the corresponding granges.
}
\description{
Combining matching transcripts
}
\details{
This is an internal function used to merge all overlapping transcripts of a breakpoint into one vector.
}
