% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{.constrict}
\alias{.constrict}
\title{constrict}
\usage{
.constrict(gr, ref = NULL, position = "middle")
}
\arguments{
\item{gr}{GRanges object}

\item{ref}{reference}

\item{position}{only 'middle' position is accepted.}
}
\value{
A constricted GRanges object.
}
\description{
constrict
}
