% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{.isSymbolic}
\alias{.isSymbolic}
\title{Determining whether the variant is a symbolic allele.}
\usage{
.isSymbolic(r, a)
}
\arguments{
\item{r}{Reference vector.}

\item{a}{ALT vector.}
}
\value{
A logical list of which the length is the same with the input object.
}
\description{
Determining whether the variant is a symbolic allele.
}
