% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.pairwiseLCPrefix}
\alias{.pairwiseLCPrefix}
\title{vectorised pairwise longest common prefix
Returns the length of the longest common prefix for
each string pair}
\usage{
.pairwiseLCPrefix(s1, s2, ignore.case = FALSE)
}
\arguments{
\item{s1, s2}{A pair of strings.}

\item{ignore.case}{Whether cases in the strings should be ignored.}
}
\value{
The length of the longest common prefix for each string pair.
}
\description{
vectorised pairwise longest common prefix
Returns the length of the longest common prefix for
each string pair
}
