% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDetect.R
\name{.scoreByTranscripts}
\alias{.scoreByTranscripts}
\title{Ranking matching transcripts}
\usage{
.scoreByTranscripts(genes, transcripts.col)
}
\arguments{
\item{genes}{TxDb object of genes. hg19 and hg38 are supported in the current version.}

\item{transcripts.col}{A vector of transcript names.}
}
\value{
A dataframe with two columns, tx_name and score.
}
\description{
Ranking matching transcripts
}
\details{
This is an internal function which returns overlapping transcript names with ranking scores. 
The ranking score is the proportion of exon-exon fusions (intronic deletion events) detected for a given transcript.
}
