% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{.svLen}
\alias{.svLen}
\title{Returns the structural variant length of the first allele}
\usage{
.svLen(vcf)
}
\arguments{
\item{vcf}{VCF object}
}
\value{
Structural variant lengths of the first allele.
}
\description{
Returns the structural variant length of the first allele
}
