% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittesthelper.R
\name{.testrecord}
\alias{.testrecord}
\title{loads a VCF containing the given records}
\usage{
.testrecord(record)
}
\arguments{
\item{record}{string vector of record to write}
}
\description{
loads a VCF containing the given records
}
