% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{.toVcfBreakendNotationAlt}
\alias{.toVcfBreakendNotationAlt}
\title{Converts to breakend notation}
\usage{
.toVcfBreakendNotationAlt(gr, insSeq = gr$insSeq, ref = gr$REF)
}
\arguments{
\item{gr}{GRanges object.}

\item{insSeq}{insert sequence of the GRanges.}

\item{ref}{reference sequence of the GRanges.}
}
\value{
breakendAlt or breakpointAlt depending on whether the variant is partnered.
}
\description{
Converts to breakend notation
}
