% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{elementExtract}
\alias{elementExtract}
\title{Extracts the element of each element at the given position}
\usage{
elementExtract(x, offset = 1)
}
\arguments{
\item{x}{list-like object}

\item{offset}{offset of list}
}
\description{
Extracts the element of each element at the given position
}

