% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insSeq.R
\name{numtDetect_insseq}
\alias{numtDetect_insseq}
\title{Detecting nuclear mitochondria fusion events from unmapped insertion sequences.}
\usage{
numtDetect_insseq(gr, genomeMT, min_len = 20, min.Align = 0.8)
}
\arguments{
\item{gr}{A GRanges object}

\item{genomeMT}{A genome object of the mitochondria.}

\item{min_len}{The minimum length allowed of the insertion sequences.
Default value is 20.}

\item{min.Align}{The minimum alignment score allowed between the insertion 
sequence and MT genome.}
}
\value{
A nested list of GRanges objects of candidate NUMTs.
}
\description{
Detecting nuclear mitochondria fusion events from unmapped insertion sequences.
}
\details{
This function looks for NUMTs which the insertion MT sequences come from 
insertion sequences reported by SV callers.
}
\keyword{internal}
