% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTbp.R
\name{numtDetect_MT}
\alias{numtDetect_MT}
\title{Detecting nuclear mitochondria fusion events from breakpoints connected to MT reference genome.}
\usage{
numtDetect_MT(gr, max_ins_dist = 10)
}
\arguments{
\item{gr}{A GRanges object}

\item{max_ins_dist}{The maximum distance allowed on the reference genome between the paired insertion sites.
Only intra-chromosomal NUMT events are supported. Default value is 10.}
}
\value{
A nested list of GRanges objects of candidate NUMTs.
}
\description{
Detecting nuclear mitochondria fusion events from breakpoints connected to MT reference genome.
}
\details{
This function looks for NUMTs which the insertion MT sequences come from known NUMT sites.
}
\keyword{internal}
