% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insSeq.R
\name{seqAlignment.score}
\alias{seqAlignment.score}
\title{Calculating the alignment score between a DNA sequence and target genome.}
\usage{
seqAlignment.score(seq, genome)
}
\arguments{
\item{seq}{A string of DNA sequence.}

\item{genome}{An XString of the target genome.}
}
\value{
A alignment score between a DNA sequence and target genome.
}
\description{
Calculating the alignment score between a DNA sequence and target genome.
}
\details{
This function calculates the alignment score between a DNA sequence and target genome.
}
\keyword{internal}
