% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{simpleEventLength}
\alias{simpleEventLength}
\title{Length of event if interpreted as an isolated breakpoint.}
\usage{
simpleEventLength(gr)
}
\arguments{
\item{gr}{breakpoint GRanges object}
}
\value{
Length of the simplest explaination of this breakpoint/breakend.
}
\description{
Length of event if interpreted as an isolated breakpoint.
}
