% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralVariantAnnotation.R
\docType{package}
\name{StructuralVariantAnnotation}
\alias{StructuralVariantAnnotation}
\alias{StructuralVariantAnnotation-package}
\title{StructuralVariantAnnotation: a package for VCF}
\description{
plyranges is a dplyr like API to the Ranges/GenomicRanges infrastructure
in Bioconductor.
}
\details{
StructuralVariantAnnotation contains useful helper functions for reading
and interpreting structural variants calls. The packages contains functions
for parsing VCFs from a number of popular caller as well as functions for
dealing with breakpoints involving two separate genomic loci. The package
supports encoding breakpoints as `GRanges` objects and BEDPE-formatted
data frame.

* Parse VCF objects with the `breakpointRanges()` function.
   * Find breakpoint overlaps with the `findBreakpointOverlaps()`
  and `countBreakpointOverlaps()` functions.
   * Generate BEDPE files for circos plot with `breakpointgr2bedpe()` function.
   * ...

For more details on the features of StructuralVariantAnnotation, read the vignette:
`browseVignettes(package = "StructuralVariantAnnotation")`
}
