% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svaNUMT.R
\docType{package}
\name{svaNUMT}
\alias{svaNUMT}
\title{svaNUMT: a package for NUMT detection}
\description{
svaNUMT contains functions for detecting NUMT events from structural variant 
calls. svaNUMT contains functions for detecting NUMT events from structural 
variant calls. It takes structural variant calls in GRanges of breakend 
notation and identifies NUMTs by nuclear-mitochondrial breakend junctions. 
The main function reports candidate NUMTs if there is a pair of valid 
insertion sites found on the nuclear genome within a certain distance 
threshold. The candidate NUMTs are reported by events.
}
\details{
For more details on the features of StructuralVariantAnnotation, read the vignette:
`browseVignettes(package = "svaNUMT")`
}
