% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svqsc.R
\name{svqsc_generate_model}
\alias{svqsc_generate_model}
\title{generates a calibrated model from a numeric data frame with a tp field}
\usage{
svqsc_generate_model(trainingdf, ...)
}
\description{
generates a calibrated model from a numeric data frame with a tp field
}

