% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svqsc.R
\name{svqsc_score}
\alias{svqsc_score}
\title{Score the given variants according to the given model}
\usage{
svqsc_score(df, model)
}
\arguments{
\item{df}{breakpoint annotations}

\item{model}{scoring model generated from \code{\link{svqsc_train}}}
}
\description{
Score the given variants according to the given model
}

