% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svqsc.R
\name{svqsc_train}
\alias{svqsc_train}
\title{Generates a calibrated variant scoring model}
\usage{
svqsc_train(traininggr, trainingdf, truthgr, countAllMatchingCalls = FALSE,
  requiredSupportingReads = 1, ...)
}
\arguments{
\item{x}{list-like object}

\item{offset}{offset of list}
}
\value{
Variant scoring model
}
\description{
Generates a calibrated variant scoring model
}

