% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-DataFrame.R
\name{transformFunctionsFor}
\alias{transformFunctionsFor}
\title{Returns a list of transformation functions for the given variant caller}
\usage{
transformFunctionsFor(caller)
}
\description{
Returns a list of transformation functions for the given variant caller
}

