% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-DataFrame.R
\name{unpack}
\alias{unpack}
\title{Expands any columns with multiple values (such as Lists) into separate columns}
\usage{
unpack(x, ...)
}
\description{
Expands any columns with multiple values (such as Lists) into separate columns
}

