% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{align_breakpoints}
\alias{align_breakpoints}
\title{Adjusting the nominal position of a pair of partnered breakpoint.}
\usage{
align_breakpoints(
  vcf,
  align = c("centre"),
  is_higher_breakend = names(vcf) < info(vcf)$PARID
)
}
\arguments{
\item{vcf}{A VCF object.}

\item{align}{The alignment type.}

\item{is_higher_breakend}{Breakpoint ID ordering.}
}
\value{
A VCF object with adjusted nominal positions.
}
\description{
Adjusting the nominal position of a pair of partnered breakpoint.
}
