% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R, R/plotting.R
\name{bedpe2breakpointgr}
\alias{bedpe2breakpointgr}
\alias{breakpointgr2bedpe}
\title{Loading a breakpoint GRanges from a BEDPE file}
\usage{
bedpe2breakpointgr(file, placeholderName = "bedpe")

breakpointgr2bedpe(gr)
}
\arguments{
\item{file}{A BEDPE file. See \url{https://bedtools.readthedocs.io/en/latest/content/general-usage.html} for details.}

\item{placeholderName}{Prefix to use to ensure each entry has a unique ID.}

\item{gr}{A GRanges object.}
}
\value{
Breakpoint GRanges object.

A BEDPE-formatted data frame.
}
\description{
Loading a breakpoint GRanges from a BEDPE file

Converting breakpoint GRanges to BEDPE-like dataframe
}
\details{
A BEDPE file is taken as an input and converted to a breakpoint GRanges object. The file is first loaded using
rtracklayer::import as a Pairs object, then converted to GRanges.

\code{breakpointgr2bedpe} converts a breakpoint GRanges to a BEDPE-formatted
dataframe. The BEDPE format consists of two sets of genomic loci, optinal
columns of name, score, strand1, strand2 and any user-defined fields.
See \url{https://bedtools.readthedocs.io/en/latest/content/general-usage.html}
for more details of BEDPE format.
}
\examples{
bedpe.file <- system.file("extdata", "gridss.bedpe", package = "StructuralVariantAnnotation")
bedpe2breakpointgr(bedpe.file)
bedpe2breakpointgr(bedpe.file, placeholderName='gridss')
#coverting a GRanges object to BEDPE-like dataframe
vcf.file <- system.file("extdata", "gridss.vcf", package = "StructuralVariantAnnotation")
vcf <- VariantAnnotation::readVcf(vcf.file, "hg19")
gr <- breakpointRanges(vcf)
breakpointgr2bedpe(gr)
}
