% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{breakpointRanges}
\alias{breakpointRanges}
\title{Extracts the structural variants as a BreakendGRanges}
\usage{
breakpointRanges(x, ...)
}
\description{
Extracts the structural variants as a BreakendGRanges
}
\details{
Structural variants are converted to breakend notation.

Due to ambiguities in the VCF specifications, structural variants
with multiple alt alleles are not supported.

If HOMLEN or HOMSEQ is defined without CIPOS, it is assumed that
the variant position is left aligned.
}

