% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{countBreakpointOverlaps}
\alias{countBreakpointOverlaps}
\title{Finds common breakpoints between the two breakpoint sets}
\usage{
countBreakpointOverlaps(querygr, subjectgr, countOnlyBest = FALSE,
  breakpointScoreColumn = "QUAL", maxgap=-1L, minoverlap=0L,
  ignore.strand = FALSE, sizemargin = 0.25,
  restrictMarginToSizeMultiple = 0.5)
}
\arguments{
\item{countOnlyBest}{count each subject breakpoint as overlaping only the
best overlapping query breakpoint.}

\item{breakpointScoreColumn}{query column defining a score for
determining which query breakpoint is considered the best when countOnlyBest=TRUE}
}
\value{
an integer vector containing the tabulated query overlap hits
}
\description{
Finds common breakpoints between the two breakpoint sets
}
\details{
See GenomicRanges::countOverlaps-methods
}

