% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittesthelper.R
\name{.testfile}
\alias{.testfile}
\title{testthat helper utility to locate files used
for package tests}
\usage{
.testfile(filename, location = "extdata")
}
\arguments{
\item{filename}{Name of the test file.}

\item{location}{Directory of the test file.}
}
\description{
testthat helper utility to locate files used
for package tests
}
