% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{.toVcfBreakendNotationAlt}
\alias{.toVcfBreakendNotationAlt}
\title{Converts to breakend notation}
\usage{
.toVcfBreakendNotationAlt(gr, insSeq = gr$insSeq, ref = gr$REF)
}
\description{
Converts to breakend notation
}
