% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{findBreakpointOverlaps}
\alias{findBreakpointOverlaps}
\title{Finds overlapping breakpoints by requiring that breakends on
boths sides overlap}
\usage{
findBreakpointOverlaps(query, subject, maxgap = 0L, minoverlap = 1L,
  ignore.strand = FALSE)
}
\description{
Finds overlapping breakpoints by requiring that breakends on
boths sides overlap
}
\details{
See GenomicRanges::findOverlaps-methods for details of overlap calculation
}

