% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions-VCF.R
\name{isStructural}
\alias{isStructural}
\title{Determines whether the variant is a structural variant}
\usage{
isStructural(x, ...)
}
\description{
Determines whether the variant is a structural variant
}

