% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{pairs2breakpointgr}
\alias{pairs2breakpointgr}
\title{Converts a BEDPE Pairs containing pairs of GRanges loaded using rtracklayer::import to a breakpointgr}
\usage{
pairs2breakpointgr(pairs, placeholderName = "bedpe")
}
\arguments{
\item{pairs}{pairs object}

\item{placeholderName}{prefix to use to ensure ids are unique}
}
\value{
breakpoint GRanges object
}
\description{
Converts a BEDPE Pairs containing pairs of GRanges loaded using rtracklayer::import to a breakpointgr
}
