% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakpointGRanges.R
\name{partner}
\alias{partner}
\title{GRanges representing the breakend coordinates of
structural variants
#@export
Partner breakend for each breakend.}
\usage{
partner(gr)
}
\description{
GRanges representing the breakend coordinates of
structural variants
#@export
Partner breakend for each breakend.
}
\details{
All breakends must have their partner breakend included
in the GRanges.
}

