% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralVariantAnnotation.R
\docType{package}
\name{StructuralVariantAnnotation}
\alias{StructuralVariantAnnotation}
\alias{StructuralVariantAnnotation-package}
\title{StructuralVariantAnnotation: a package for SV annotation}
\description{
StructuralVariantAnnotation contains useful helper functions for reading
and interpreting structural variants calls. The packages contains functions
for parsing VCFs from a number of popular caller as well as functions for
dealing with breakpoints involving two separate genomic loci. The package
takes a `GRanges` based breakend-centric approach.
}
\details{
* Parse VCF objects with the `breakpointRanges()` and `breakendRanges()`functions.
   * Find breakpoint overlaps with the `findBreakpointOverlaps()`
  and `countBreakpointOverlaps()` functions.
   * Generate BEDPE files for circos plot with `breakpointgr2pairs()` function.
   * ...

For more details on the features of StructuralVariantAnnotation, read the vignette:
`browseVignettes(package = "StructuralVariantAnnotation")`
}
